Sources needed for API moved to commons module.
