+++ commons-app/src/main/java/cz/cuni/mff/xrg/odcs/commons/app/dpu/validator/DataUnitNamesValidator.java
@@ -1,17 +1,16 @@
-package cz.cuni.mff.xrg.odcs.frontend.dpu.validator;
-
-import java.util.HashSet;
-import java.util.List;
-
-import org.springframework.beans.factory.annotation.Autowired;
-import org.springframework.stereotype.Component;
+package cz.cuni.mff.xrg.odcs.commons.app.dpu.validator;
 
 import cz.cuni.mff.xrg.odcs.commons.app.data.DataUnitDescription;
 import cz.cuni.mff.xrg.odcs.commons.app.dpu.DPUExplorer;
 import cz.cuni.mff.xrg.odcs.commons.app.dpu.DPUTemplateRecord;
 import cz.cuni.mff.xrg.odcs.commons.app.module.DPUValidator;
 import cz.cuni.mff.xrg.odcs.commons.app.module.DPUValidatorException;
-import cz.cuni.mff.xrg.odcs.frontend.i18n.Messages;
+import cz.cuni.mff.xrg.odcs.commons.app.i18n.Messages;
+import org.springframework.beans.factory.annotation.Autowired;
+import org.springframework.stereotype.Component;
+
+import java.util.HashSet;
+import java.util.List;
 
 /**
  * Check for duplicity in names of input and output data units.
@@ -35,7 +34,7 @@ class DataUnitNamesValidator implements DPUValidator {
      * exception.
      * 
      * @param dataUnits
-     * @throws DPUValidatorException
+     * @throws cz.cuni.mff.xrg.odcs.commons.app.module.DPUValidatorException
      */
     private void check(List<DataUnitDescription> dataUnits) throws DPUValidatorException {
         HashSet<String> names = new HashSet<>();
